requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
logger==1.4
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
